+++ DotML.Cli/src/Logging/BaseOutputLogger.cs
@@ -13,63 +13,63 @@ public abstract class BaseOutputLogger : IOutputLogger {
     }
 
 
-    public virtual void Visit(ConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(ConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(DepthwiseConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(DepthwiseConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(TransposeConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(TransposeConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(PixelShuffle layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(PixelShuffle layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(PoolingLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(PoolingLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(FlatteningLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(FlatteningLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(DropoutLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(DropoutLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(LayerNorm layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(LayerNorm layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(BatchNorm layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(BatchNorm layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(DenseLinearLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(DenseLinearLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(ActivationLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(ActivationLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(SoftmaxLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(SoftmaxLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(InputCapture capture, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(InputCapture capture, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(AdditionSkipConnection capture, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(AdditionSkipConnection capture, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public virtual void Visit(ConcatenationSkipConnection capture, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public virtual void Visit(ConcatenationSkipConnection capture, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 }
