+++ DotML.Cli/src/Logging/ImageOutputLogger.cs
@@ -6,7 +6,7 @@ namespace DotML.Cli.Logging;
 public class ImagesOutputLogger : BaseOutputLogger {
     public ImagesOutputLogger(DirectoryInfo logDir) : base(logDir) { }
 
-    private void EmitImage(string layerName, BatchedFeatureSet<double> args) {
+    private void EmitImage(string layerName, BatchedFeatureSet<float> args) {
         for (var batch = 0; batch < args.Batches; batch++) {
             var features = args[batch];
             var dir_path = Path.Combine(LogDirectory.FullName, $"Batch-{batch}");
@@ -36,76 +36,76 @@ public class ImagesOutputLogger : BaseOutputLogger {
         }
     }
 
-    public override void Visit(ConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(ConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(ConvolutionLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(DepthwiseConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(DepthwiseConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(DepthwiseConvolutionLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(TransposeConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(TransposeConvolutionLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(TransposeConvolutionLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(PixelShuffle layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(PixelShuffle layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(PixelShuffle)}", args.Output);
         return;
     }
 
-    public override void Visit(PoolingLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(PoolingLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(PoolingLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(FlatteningLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(FlatteningLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(FlatteningLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(DropoutLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(DropoutLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(DropoutLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(LayerNorm layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(LayerNorm layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(LayerNorm)}", args.Output);
         return;
     }
 
-    public override void Visit(BatchNorm layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(BatchNorm layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(BatchNorm)}", args.Output);
         return;
     }
 
-    public override void Visit(DenseLinearLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(DenseLinearLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(DenseLinearLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(ActivationLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(ActivationLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(ActivationLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(SoftmaxLayer layer, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(SoftmaxLayer layer, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(SoftmaxLayer)}", args.Output);
         return;
     }
 
-    public override void Visit(InputCapture capture, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(InputCapture capture, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         return;
     }
 
-    public override void Visit(AdditionSkipConnection capture, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(AdditionSkipConnection capture, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(AdditionSkipConnection)}", args.Output);
         return;
     }
 
-    public override void Visit(ConcatenationSkipConnection capture, (int LayerIndex, BatchedFeatureSet<double> Output) args) {
+    public override void Visit(ConcatenationSkipConnection capture, (int LayerIndex, BatchedFeatureSet<float> Output) args) {
         EmitImage($"Layer-{args.LayerIndex} {nameof(ConcatenationSkipConnection)}", args.Output);
         return;
     }
